+++ src/dog_marker/api/v1/endpoints/user_entries.py
@@ -19,12 +19,12 @@ async def get_user_entries(
     page_info: Pagination = Depends(query_pagination),
     entry_service: EntryService = Depends(get_service(EntryService)),
 ) -> Iterable[EntrySchema]:
-    new_entry = entry_service.all(
+    entries = entry_service.get_all_by_owner(
         page_info=page_info,
         owner_id=user_id,
         warning_level=warning_level,
     )
-    return new_entry
+    return entries
 
 
 @router.post(
@@ -54,7 +54,7 @@ async def put_entry(
     update_entry: UpdateEntrySchema,
     entry_service: EntryService = Depends(get_service(EntryService)),
 ) -> EntrySchema:
-    updated_entry = entry_service.update_entry(entry_id, user_id, update_entry)
+    updated_entry = entry_service.update(entry_id, user_id, update_entry)
 
     return updated_entry
 
@@ -64,3 +64,23 @@ async def delete_entry_for_user(
     user_id: UUID, entry_id: UUID, entry_service: EntryService = Depends(get_service(EntryService))
 ) -> None:
     entry_service.delete(entry_id, user_id)
+
+
+@router.get("/{user_id}/entries/trash", response_model=list[EntrySchema], operation_id="get_trashed_user_entries")
+async def get_trashed_entries(
+    user_id: UUID,
+    page_info: Pagination = Depends(query_pagination),
+    entry_service: EntryService = Depends(get_service(EntryService)),
+) -> Iterable:
+    entries = entry_service.deleted_entries(page_info=page_info, user_id=user_id)
+    return entries
+
+
+@router.post("/{user_id}/entries/trash/{entry_id}/undo", response_model=EntrySchema, operation_id="undo_trashed_entry")
+async def post_undo_deleted_entry(
+    user_id: UUID,
+    entry_id: UUID,
+    entry_service: EntryService = Depends(get_service(EntryService)),
+) -> Iterable:
+    entries = entry_service.undo_deleted_entry(entry_id=entry_id, user_id=user_id)
+    return entries